requirements.txt
colorama; sys_platform=='win32'
h5py<3.0
jsonschema
keras==2.2.4
numexpr
numpy>=1.18,!=1.19.4
scipy
tensorflow_gpu==1.13.1
tifffile
tqdm>=4.41.0
